.env checks + modify routes to start with /api
Some small tweaks, taking into consideration that ts is on github now made some changes to account for user error.

app.py on run checks if .env exists if it doesnt an errors returned. If .env exists but theres no value in GROQ_KEY then an error is returned. These errors are displayed on the page and no api requests can be sent to flask.

Changed routes to start with /api to allow for blocking requests on error.