Allow for the use of local models
Massive change here. Spent many hours straight attempting to implement the choice between online and locally hosted models and got that shi workin.

Flask reads .env to find which backend the users trying to use. If ollama the backend calls out to ollama at the url provided again in .env and if backend if groq it calls out to groq api using the key.

Storage and memory functionality stayed kinda the same.. it was a lot worse getting memory working and i havent tested on other models yet so it may bug for other models who knows.

markdown files created for local and online setup as well as expanding how-it-works to explain local side.